+++ Tuning/FFTharmonics2.py
@@ -1,7 +1,6 @@
 import logging
-from math import gcd
+import math
 from operator import itemgetter
-from typing import Any
 
 from numpy import sqrt, mean, append, array
 from numpy.typing import NDArray
@@ -9,15 +8,57 @@ from numpy.typing import NDArray
 # internal
 import parameters
 from FFTaux import mytimer
-# from minimize_bruteforce import final_fit
-from minimize_SLSQP import final_fit
+from minimize_SLSQP_class import MinimizeSLSQP
 
 I_MAX = int(16_000 / parameters.FREQUENCY_LOWER)
 
 
+def l2min_new(
+        ind: list,
+        f0: float,
+        b: float,
+) -> float:
+    """
+    Note: L2 for testing and reference, not currently used in minimization.
+    returns the cost function for a regression on the L2 norm
+    l2 = sum( ( ( f_i(measured) - f_i(calculated) ) / f_i(measured) )**2 )
+    where f_i(measured) is considered to its nearest neighbor either
+    f_i(calculated) or f_i+1(calculated)
+    :param ind: list - measured resonance frequencies as from peaks (FFT)
+    after being cleansed, duplicates removed, etc.
+    :param f0: float - base frequency
+    :param b: float - inharmonicity, such that f = i * f0 * sqrt(1. + b * i**2)
+    :return: float - l2 cost function
+    """
+    l2 = 0.  # l2 cost function
+    j = 1
+    # loop over peaks found
+    for found in ind:
+        fl = j * f0 * sqrt(1. + b * j ** 2)
+        for i in range(j, I_MAX):
+            fh = (i + 1) * f0 * sqrt(1. + b * (i + 1) ** 2)
+            j = i
+            if found < fl and i == 1:
+                diff = fl - found
+                l2 += diff * diff / found / found
+                break
+            elif fl <= found < fh:
+                if (found - fl) < (fh - found):
+                    diff = found - fl
+                else:
+                    diff = fh - found
+                l2 += diff * diff / found / found
+                break
+            else:
+                fl = fh
+
+    return l2
+
+
 def l1min_new(
         ind: list,
-        x0: list
+        f0: float,
+        b: float,
 ) -> float:
     """
     returns the cost function for a regression on the L1 norm
@@ -26,16 +67,17 @@ def l1min_new(
     f_i(calculated) or f_i+1(calculated)
     :param ind: list - measured resonance frequencies as from peaks (FFT)
     after being cleansed, duplicates removed, etc.
-    :param x0: list - [f0, b] such that f = i * x0[0] * sqrt(1. + x0[1] * i**2)
+    :param f0: float - base frequency
+    :param b: float - inharmonicity, such that f = i * f0 * sqrt(1. + b * i**2)
     :return: float - l1 cost function
     """
     l1 = 0.  # l1 cost function
     j = 1
     # loop over peaks found
     for found in ind:
-        fl = j * x0[0] * sqrt(1. + x0[1] * j ** 2)
+        fl = j * f0 * sqrt(1. + b * j ** 2)
         for i in range(j, I_MAX):
-            fh = (i + 1) * x0[0] * sqrt(1. + x0[1] * (i + 1) ** 2)
+            fh = (i + 1) * f0 * sqrt(1. + b * (i + 1) ** 2)
             j = i
             if found < fl and i == 1:
                 l1 += (fl - found) / found
@@ -80,8 +122,7 @@ def harmonics(peaks: list[tuple]) -> list:
         positions of first NPARTIAL partials
     """
     initial = list()
-    l1: dict[int, list[float]] = dict()
-    l1_mean: dict[int, Any] = dict()
+    l1: dict[tuple[int, float], list[float]] = dict()
     f_n = list()
 
     # sort by frequency asc. and make list of indices (positions) and heights
@@ -91,12 +132,14 @@ def harmonics(peaks: list[tuple]) -> list:
     logging.debug("ind: " + str(ind))
     logging.debug("height: " + str(height))
 
-    # loop through all combinations of partials up to NPARTIAL
-    for m in range(1, parameters.NPARTIAL):
-        for k in range(m + 1, parameters.NPARTIAL):
-            # loop through all peaks found (ascending, nested loops)
-            for i in range(0, len(ind)):
-                for j in range(i + 1, len(ind)):
+    nex = 1
+    # loop through all peaks found (ascending, nested loops)
+    for i in range(0, len(ind)):
+        for j in range(i + 1, len(ind)):
+            # loop through all combinations of partials up to NPARTIAL
+            for m in range(nex, parameters.NPARTIAL):
+                for k in range(m + 1, parameters.NPARTIAL):
+                    # calculate inharmonicity factor b from two peaks ind[i], ind[j]
                     tmp = ((ind[j] * m) / (ind[i] * k)) ** 2
                     try:
                         b = (tmp - 1.) / (k ** 2 - tmp * m ** 2)
@@ -108,11 +151,12 @@ def harmonics(peaks: list[tuple]) -> list:
                     if -0.0001 < b < parameters.INHARM:
                         # allow also negative b value > -0.0001 for
                         # uncertainties in the line fitting
+                        # calculate fundamental frequency from lower partial
                         f_fundamental = ind[i] / (m * sqrt(1. + b * m ** 2))
                         if not (parameters.FREQUENCY_LOWER
                                 < f_fundamental
                                 < parameters.FREQUENCY_UPPER):
-                            break  # break two loops here
+                            break
                         element = [
                             m, k, ind[i], ind[j], max(b, 0.), f_fundamental
                         ]  # always b >= 0
@@ -120,80 +164,63 @@ def harmonics(peaks: list[tuple]) -> list:
                             if (element[3] == initial[-1][3]
                                     and element[0] == initial[-1][0]):
                                 # remove previous doublette on upper frequency and lower partial
+                                # print("removed doublette", initial[-1], element)
                                 initial.pop()
                         initial.append(element)
-                        break  # break two loops here
-                break
+        nex += 1  # increase lower partial for next higher peak found
 
-    for item in initial:
-        # prepare for two partials with no common divisor open an empty list
-        if gcd(item[0], item[1]) == 1:
-            if item[0] not in l1:
-                # create dict keys with empty list values
-                l1[item[0]] = list()
-        logging.debug(
-            "partials: {0:2d} {1:2d} lower: {2:10.4f} upper: {3:10.4f} "
-            "B: {4: .1e} fundamental: {5:10.4f}".format(*item))
-
-    """
-    disregard fundamentals for records, where both partials have a common 
-    divisor (gcd). Consider all fundamentals with no common divisor only.
-    """
     if initial:
-        av = array([])
-        selected = array([])
-        no_of_peak_combi = 0
+        l1_min = float('inf')
 
-        if len(l1) > 1:
-            # if more than one lower partials with gcd=1
-            for key in l1:
-                for dat in filter(lambda x: x[0] == key, initial):
-                    # Add all l1 values to list for same lower partial
-                    t_new = l1min_new(ind=ind, x0=[dat[5], dat[4]])
-                    l1[key].append(t_new)
-                # l1 cost function averaged for equal lower partials
-                l1_mean[key] = mean(l1[key])
-            # identify lower partial with minimum l1
-            selected = array(
-                list(
-                    filter(lambda x: x[0] == min(l1_mean, key=l1_mean.get), initial))
-            )
-            av = selected.mean(axis=0)
-            no_of_peak_combi = selected.shape[0]
-            logging.debug("L1: {}".format(l1_mean))
-        elif len(l1) == 1:
-            # if only one lower partial with gcd=1
-            selected = array(
-                list(
-                    filter(lambda x: x[0] == list(l1.keys())[0], initial))
-            )
-            av = selected.mean(axis=0)
-            no_of_peak_combi = selected.shape[0]
+        for item in initial:  # if found any partial combinations
+            t = (item[0], item[2])
+            initial_log = [
+                math.log(item[4]) if item[4] > 0.0 else float('nan'),
+                item[5]
+            ]
+            if t not in l1:
+                l1[t] = list()
+            l1[t].append(initial_log)  # type: ignore
+            print(l1)
+            logging.debug(
+                "partials: {0:2d} {1:2d} lower: {2:10.4f} upper: {3:10.4f} "
+                "B: {4: .1e} fundamental: {5:10.4f}".format(*item))
 
-        if av.size == 0:
-            # if no gcd=1 found, take first entry for the lowest partial
-            av = array(list(
-                filter(lambda x: x[0] == initial[0][0], initial))
-            ).mean(axis=0)
+        for key, val in l1.items():
+            arrays = [array(x) for x in val]
+            initial_av = [mean(k) for k in zip(*arrays)]
+            t_new = l1min_new(
+                ind=ind,
+                f0=float(initial_av[1]),
+                b=math.exp(initial_av[0]) if not math.isnan(initial_av[0]) else 0.)
+            if t_new < l1_min:
+                l1_min = t_new
+                logging.debug("Last L1 minimum: {}".format(l1_min))
+                selected = array(
+                    list(
+                        filter(lambda x: (x[0], x[2]) == key, initial))
+                )
+                base_frequency = float(initial_av[1])
+                inharmonicity = math.exp(initial_av[0]) if not math.isnan(initial_av[0]) else 0.
 
-        base_frequency = av[5]
-        inharmonicity = av[4]
         if (parameters.FREQUENCY_LOWER
                 < base_frequency
                 < parameters.FREQUENCY_UPPER):
-            if no_of_peak_combi > 1:
-                identified = select_list(selected=selected)
-                base_frequency, inharmonicity = final_fit(
-                    av=av,
-                    ind=identified
-                )
-                logging.debug(
-                    "Initial: f_0 = {0:.3f} Hz, B = {1:.3e} "
-                    "Final: f_0 = {2:.3f} Hz, B = {3:.3e}".format(
-                        av[5], av[4], base_frequency, inharmonicity)
-                )
-            for n in range(1, I_MAX):
-                # for n in range(1, parameters.NPARTIAL):
+            identified = select_list(selected=selected)
+            base_frequency_final, inharmonicity_final = (
+                MinimizeSLSQP(norm=parameters.COST_FUNCTION)(
+                    ind=identified,
+                    f0=base_frequency,
+                    b=inharmonicity
+                ))
+            logging.debug(
+                "initial: f_0 = {0:.3f} Hz, B = {1:.3e} "
+                "Final: f_0 = {2:.3f} Hz, B = {3:.3e}".format(
+                    base_frequency, inharmonicity,
+                    base_frequency_final, inharmonicity_final)
+            )
+            # display synthetic spectrum
+            for n in range(1, parameters.NPARTIAL):
                 f_synth = base_frequency * n * sqrt(
                     1. + inharmonicity * n ** 2)
                 if f_synth < 12_000:
@@ -213,7 +240,8 @@ def harmonics(peaks: list[tuple]) -> list:
         if parameters.FREQUENCY_LOWER < f1 < parameters.FREQUENCY_UPPER:
             f_n.append(f1)
             logging.info(
-                "Best result: f_1 = {0:.2f} Hz, B = {1:.1e}".format(f1, 0.)
+                "Best result from strongest line: f_1 = {0:.2f} Hz, B = {1:.1e}"
+                .format(f1, 0.)
             )
 
     return f_n